refactor as per review
